Create Front-End (React):  Set up a new React project using a tool like Create React App:
Create the user interface components for the application, including a search bar and a display area for weather information.

Implement the logic to send HTTP requests to the back end when a user searches for a location.

Display the weather data (current conditions, temperature, humidity) on the front end.

Ensure that the user interface is responsive and works on both desktop and mobile devices. You can use responsive design libraries like Bootstrap or Material-UI to help with this.